LPS-167967 Move ObjectFieldUtil to package com.liferay.object.field.util
