+++ modules/apps/object/object-api/src/main/java/com/liferay/object/field/util/ObjectFieldUtil.java
@@ -12,11 +12,12 @@
  * details.
  */
 
-package com.liferay.object.util;
+package com.liferay.object.field.util;
 
 import com.liferay.object.model.ObjectField;
 import com.liferay.object.model.ObjectFieldSetting;
 import com.liferay.object.service.ObjectFieldLocalServiceUtil;
+import com.liferay.object.util.LocalizedMapUtil;
 
 import java.util.Collections;
 import java.util.List;
